requirements/requirements-base.txt
djangorestframework-jwt>=1.11.0
six>=1.10.0
